+++ src/main/java/de/canitzp/solarhelmet/SolarHelmetData.java
@@ -2,6 +2,7 @@ package de.canitzp.solarhelmet;
 
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.data.DataGenerator;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.recipes.*;
@@ -14,6 +15,7 @@ import net.neoforged.fml.common.Mod;
 import net.neoforged.neoforge.client.model.generators.ItemModelProvider;
 import net.neoforged.neoforge.common.Tags;
 import net.neoforged.neoforge.common.data.ExistingFileHelper;
+import net.neoforged.neoforge.common.data.internal.NeoForgeRecipeProvider;
 import net.neoforged.neoforge.data.event.GatherDataEvent;
 import org.jetbrains.annotations.NotNull;
 
@@ -29,7 +31,7 @@ public class SolarHelmetData {
         CompletableFuture<HolderLookup.Provider> lookupProvider = event.getLookupProvider();
 
         generator.addProvider(event.includeClient(), new ItemModel(generator.getPackOutput(), helper));
-        generator.addProvider(event.includeServer(), new Recipe(generator.getPackOutput(), lookupProvider));
+        event.createProvider(event.includeServer(), Recipe.Runner::new);
     }
 
     public static class ItemModel extends ItemModelProvider {
@@ -51,23 +53,39 @@ public class SolarHelmetData {
 
     public static class Recipe extends RecipeProvider {
 
-        public Recipe(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider) {
-            super(output, lookupProvider);
+        public Recipe(HolderLookup.Provider provider, RecipeOutput output) {
+            super(provider, output);
         }
 
         @Override
-        protected void buildRecipes(@NotNull RecipeOutput output) {
-            ShapedRecipeBuilder.shaped(RecipeCategory.TOOLS, SolarHelmet.SOLAR_MODULE_ITEM.get())
-                    .define('r', Tags.Items.STORAGE_BLOCKS_REDSTONE)
-                    .define('l', Tags.Items.GEMS_LAPIS)
-                    .define('d', Tags.Items.GEMS_DIAMOND)
-                    .define('i', Tags.Items.INGOTS_IRON)
+        protected void buildRecipes() {
+            ShapedRecipeBuilder.shaped(BuiltInRegistries.ITEM, RecipeCategory.TOOLS, SolarHelmet.SOLAR_MODULE_ITEM.get())
+                    .define('r', tag(Tags.Items.STORAGE_BLOCKS_REDSTONE))
+                    .define('l', tag(Tags.Items.GEMS_LAPIS))
+                    .define('d', tag(Tags.Items.GEMS_DIAMOND))
+                    .define('i', tag(Tags.Items.INGOTS_IRON))
                     .define('p', Items.PAPER)
                     .pattern(" r ")
                     .pattern("ldl")
                     .pattern("ipi")
                     .unlockedBy("has_diamond", has(Tags.Items.GEMS_DIAMOND))
-                    .save(output);
+                    .save(super.output);
+        }
+
+        public static final class Runner extends RecipeProvider.Runner {
+            public Runner(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider) {
+                super(output, lookupProvider);
+            }
+
+            @Override
+            protected RecipeProvider createRecipeProvider(HolderLookup.Provider lookupProvider, RecipeOutput output) {
+                return new Recipe(lookupProvider, output);
+            }
+
+            @Override
+            public String getName() {
+                return "SolarHelmet recipes";
+            }
         }
     }